LayoutTests/media/picture-in-picture/picture-in-picture-api-enter-pip-3-expected.txt

This tests that request Picture-in-Picture requires video track for the video element.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
RUN(video.src = findMediaFile("audio", "../content/test"))
EVENT(canplaythrough)
EXPECTED (error.name == 'InvalidStateError') OK
END OF TEST
